Rename weather implementation module
