+++ mcp-server/build.gradle
@@ -6,6 +6,7 @@ plugins {
 
 dependencies {
     implementation "org.springframework.ai:spring-ai-starter-mcp-server-webflux"
+    implementation project(":weather:weather-impl")
     testImplementation "org.junit.jupiter:junit-jupiter"
     testImplementation "org.springframework.boot:spring-boot-starter-test"
 }
