+++ mcp-server/src/main/java/com/cyster/mcp/McpServer.java
@@ -6,6 +6,8 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
+import com.cyster.weather.service.WeatherService;
+
 @SpringBootApplication
 public class McpServer {
     public static void main(String[] args) {
