+++ settings.gradle
@@ -22,4 +22,6 @@ include "mcp-server"
 include "chat-flux-server"
 include "mcp-tool"
 include "chat-mvp-server"
+include ":weather:weather-service"
+include ":weather:weather-impl"
