+++ weather/weather-service/src/main/java/com/cyster/weather/service/Points.java
@@ -1,0 +1,6 @@
+package com.cyster.weather.service;
+
+public record Points(Properties properties) {
+    public record Properties(String forecast) {
+    }
+}
