+++ weather/weather-service/src/main/java/com/cyster/weather/service/WeatherService.java
@@ -1,0 +1,7 @@
+package com.cyster.weather.service;
+
+public sealed interface WeatherService permits com.cyster.weather.impl.WeatherServiceImpl {
+    String getWeatherForecastByLocation(double latitude, double longitude);
+
+    String getAlerts(String state);
+}
